requirements.txt
streamlit>=1.32.0
google-genai>=0.8.0
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.3.0
Pillow>=9.0.0
plotly>=5.14.0
matplotlib>=3.7.0
seaborn>=0.12.0
nltk>=3.8.0
joblib>=1.3.0
python-dotenv>=1.0.0
requests>=2.28.0
